Rename az to alpha_zero folder Remove the unused model and trainer register Rename AZ to AlphaZero in docs
