+++ rllib/contrib/alpha_zero/doc/cartpole_plot.png
