+++ rllib/contrib/alpha_zero/models/custom_torch_models.py
@@ -3,7 +3,6 @@ from abc import ABC
 import numpy as np
 import torch
 import torch.nn as nn
-from ray.rllib.models import ModelCatalog
 from ray.rllib.models.model import restore_original_dimensions
 from ray.rllib.models.preprocessors import get_preprocessor
 from ray.rllib.models.torch.torch_modelv2 import TorchModelV2
@@ -107,6 +106,3 @@ class DenseModel(ActorCriticModel):
         self.critic_layers = nn.Sequential(
             nn.Linear(in_features=256, out_features=1))
         self._value_out = None
-
-
-ModelCatalog.register_custom_model("convnet_model", ConvNetModel)
